+++ manifest/graph.go
@@ -341,6 +341,19 @@ func (g *ModuleGraph) ChildrenOf(moduleName string) ([]*pbsubstreams.Module, err
 	return res, nil
 }
 
+func (g *ModuleGraph) HasStatefulDependencies(moduleName string) (bool, error) {
+	stores, err := g.StoresDownTo(moduleName)
+	if err != nil {
+		return false, fmt.Errorf("getting stores down to %s: %w", moduleName, err)
+	}
+
+	if len(stores) > 0 {
+		return true, nil
+	}
+
+	return false, nil
+}
+
 func (g *ModuleGraph) StoresDownTo(moduleName string) ([]*pbsubstreams.Module, error) {
 	alreadyAdded := map[string]bool{}
 	topologicalIndex := map[string]int{}